major reworking of Seq-id handling
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@32832 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/cn3d/cdd_book_ref_dialog.cpp
@@ -439,7 +439,7 @@ void CDDBookRefDialog::OnClick(wxCommandEvent& event)
                 if (&(*d) == selDescr) {
                     descrSet->Set().erase(d);
                     sSet->SetDataChanged(StructureSet::eCDDData);
-                    if (selectedItem == listbox->GetCount() - 1)
+                    if (selectedItem == (int) listbox->GetCount() - 1)
                         --selectedItem;
                     break;
                 }
